working integration test setup

+++ bindings/browser-api/ci.js
@@ -1,6 +1,13 @@
-const pkgjson = JSON.parse(fs.readFileSync("./pkg/package.json"));
-if (process.env.REF_NAME) {
-    pkgjson.version = process.env.REF_NAME
+const fs = require('fs');
+const path = require('path');
+
+const GENERATED_PACKAGEJSON_PATH = path.join(__dirname, "pkg", "package.json");
+
+if (fs.existsSync(GENERATED_PACKAGEJSON_PATH)) {
+    const pkgjson = JSON.parse(fs.readFileSync(GENERATED_PACKAGEJSON_PATH));
+    if (process.env.REF_NAME) {
+        pkgjson.version = process.env.REF_NAME
+    }
     pkgjson.name = "@easy-rpc/browser"
+    fs.writeFileSync("./pkg/package.json", JSON.stringify(pkgjson))
 }
-fs.writeFileSync("./pkg/package.json", JSON.stringify(pkgjson))
